Configure Hilt and Room database setup
- Add Hilt and KSP plugins to `app/build.gradle.kts` and include Hilt dependencies
- Create `PeopleApplication` class with `@HiltAndroidApp` and register it in `AndroidManifest.xml`
- Annotate `MainActivity` with `@AndroidEntryPoint`
- Define `AppDatabase` and `DatabaseModule` for dependency injection of Room database and `PeopleDao`
- Annotate `PeopleDao` with `@Dao`
- Refactor `PeopleEntity` and `People` model to use `Long` instead of `BigInteger` for IDs and ordering
- Downgrade Kotlin version in `libs.versions.toml` to 2.2.20 to resolve KSP compatibility issues

+++ core/database/src/main/java/com/bitmavrick/core/database/local/entity/PeopleEntity.kt
@@ -3,14 +3,13 @@ package com.bitmavrick.core.database.local.entity
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.bitmavrick.core.model.GenderType
-import java.math.BigInteger
 
 @Entity(tableName = "People")
 data class PeopleEntity (
     @PrimaryKey(autoGenerate = true)
-    val id: BigInteger,
+    val id: Long,
     val name: String,
     val age: Int,
     val gender: GenderType,
-    val orderIndex: BigInteger
+    val orderIndex: Long
 )
