+++ core/model/src/main/java/com/bitmavrick/core/model/People.kt
@@ -3,9 +3,9 @@ package com.bitmavrick.core.model
 import java.math.BigInteger
 
 data class People(
-    val id: BigInteger,
+    val id: Long,
     val name: String,
     val age: Int,
     val gender: GenderType,
-    val orderIndex: BigInteger
+    val orderIndex: Long
 )
